fix: (re)split execution path to avoid breaking changes with 0 amounts (#427)
