+++ .pre-commit-config.yaml
@@ -14,6 +14,6 @@ repos:
     - id: check-added-large-files
     - id: check-merge-conflict
 - repo: https://github.com/golangci/golangci-lint
-  rev: v1.46.2
+  rev: v1.51.0
   hooks:
     - id: golangci-lint
